refactor(types): extract shared types into dedicated types/ module
Move JSONSerializable, Message, and MessageHandler out of context/Room.tsx
into src/types/core.ts so core logic no longer depends on a React context
file. Populate the previously empty context/ and hooks/ barrel index files,
and update src/index.ts to route through those barrels. Also adds Message
and MessageHandler to the public API.
