+++ packages/react-p2p/src/context/index.ts
@@ -1,0 +1,1 @@
+export { Room, RoomContext, type RoomContextValue } from './Room';
