+++ packages/react-p2p/src/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useRoom } from './useRoom';
+export { useSharedState } from './useSharedState';
