+++ packages/react-p2p/src/types/index.ts
@@ -1,0 +1,1 @@
+export type { JSONSerializable, Message, MessageHandler } from './core';
